+++ src/roq/deribit/drop_copy.h
@@ -79,6 +79,7 @@ class DropCopy final : public core::web::Socket::Handler,
   void subscribe_portfolios(const roq::span<std::string> &currencies);
   void subscribe_changes();
   void subscribe_orders();
+  void subscribe_trades();
 
   void get_account_summary(const roq::span<std::string> &currencies);
   void get_trades(const roq::span<std::string> &currencies);
@@ -102,7 +103,8 @@ class DropCopy final : public core::web::Socket::Handler,
 
   void operator()(const server::Trace<json::Trades> &);
   void operator()(const server::Trace<json::Positions> &);
-  void operator()(const server::Trace<json::Order> &);
+  void operator()(const server::Trace<json::Order> &) override;
+  void operator()(const server::Trace<json::Trades2> &) override;
 
   void operator()(const server::Trace<json::Trade> &, bool is_last);
   void operator()(const server::Trace<json::Position> &, bool is_last);